@@ -4,14 +4,19 @@
  *--------------------------------------------------------------------------------------------*/
 
 import chalk from 'chalk';
-import { dirname, join } from 'path';
-import { rmSync } from 'fs';
-import { LOGGER, Platform, platform, Runtime } from './constants';
-import { fileGet, jsonGet } from './fetch';
-import { computeSHA256, exists, getBuildPath, unzip } from './files';
+import { dirname, join } from 'node:path';
+import { rmSync } from 'node:fs';
+import { Flavor, LOGGER, Platform, platform, Quality, Runtime } from './constants.js';
+import { fileGet, jsonGet } from './fetch.js';
+import { computeSHA256, exists, getBuildPath, unzip } from './files.js';
 
-export interface IBuild {
+export interface IBuildKind {
     readonly runtime: Runtime;
+    readonly quality: Quality;
+    readonly flavor: Flavor;
+}
+
+export interface IBuild extends IBuildKind {
     readonly commit: string;
 }
 
@@ -24,16 +29,21 @@ interface IBuildMetadata {
 
 class Builds {
 
-    async fetchBuildByVersion(runtime = Runtime.WebLocal, version: string): Promise<IBuild> {
-        const meta = await jsonGet<IBuildMetadata>(`https://update.code.visualstudio.com/api/versions/${version}.0-insider/${this.getBuildApiName(runtime)}/insider?released=true`);
+    async fetchBuildByVersion({ runtime, quality, flavor }: IBuildKind, version: string): Promise<IBuild> {
+        let meta;
+        if (quality === 'insider') {
+            meta = await jsonGet<IBuildMetadata>(`https://update.code.visualstudio.com/api/versions/${version}.0-insider/${this.getBuildApiName({ runtime, quality, flavor })}/insider?released=true`);
+        } else {
+            meta = await jsonGet<IBuildMetadata>(`https://update.code.visualstudio.com/api/versions/${version}.0/${this.getBuildApiName({ runtime, quality, flavor })}/stable?released=true`);
+        }
 
-        return { runtime, commit: meta.version };
+        return { runtime, commit: meta.version, quality, flavor };
     }
 
-    async fetchBuilds(runtime = Runtime.WebLocal, goodCommit?: string, badCommit?: string, releasedOnly?: boolean): Promise<IBuild[]> {
+    async fetchBuilds({ runtime, quality, flavor }: IBuildKind, goodCommit?: string, badCommit?: string, releasedOnly?: boolean): Promise<IBuild[]> {
 
-        // Fetch all released insider builds
-        const allBuilds = await this.fetchAllBuilds(runtime, releasedOnly);
+        // Fetch all released builds
+        const allBuilds = await this.fetchAllBuilds({ runtime, quality, flavor }, releasedOnly);
 
         let goodCommitIndex = allBuilds.length - 1;  // last build (oldest) by default
         let badCommitIndex = 0;                      // first build (newest) by default
@@ -42,9 +52,9 @@ class Builds {
             const candidateGoodCommitIndex = this.indexOf(goodCommit, allBuilds);
             if (typeof candidateGoodCommitIndex !== 'number') {
                 if (releasedOnly) {
-                    throw new Error(`Provided good commit ${chalk.green(goodCommit)} was not found in the list of insiders builds. It is either invalid or too old.`);
+                    throw new Error(`Provided good commit ${chalk.green(goodCommit)} was not found in the list of builds. It is either invalid or too old.`);
                 } else {
-                    return this.fetchBuilds(runtime, goodCommit, badCommit, true);
+                    return this.fetchBuilds({ runtime, quality, flavor }, goodCommit, badCommit, true);
                 }
             }
 
@@ -55,9 +65,9 @@ class Builds {
             const candidateBadCommitIndex = this.indexOf(badCommit, allBuilds);
             if (typeof candidateBadCommitIndex !== 'number') {
                 if (releasedOnly) {
-                    throw new Error(`Provided bad commit ${chalk.green(badCommit)} was not found in the list of insiders builds. It is either invalid or too old.`);
+                    throw new Error(`Provided bad commit ${chalk.green(badCommit)} was not found in the list of builds. It is either invalid or too old.`);
                 } else {
-                    return this.fetchBuilds(runtime, goodCommit, badCommit, true);
+                    return this.fetchBuilds({ runtime, quality, flavor }, goodCommit, badCommit, true);
                 }
             }
 
@@ -86,70 +96,70 @@ class Builds {
         return undefined;
     }
 
-    private async fetchAllBuilds(runtime: Runtime, releasedOnly = false): Promise<IBuild[]> {
-        const url = `https://update.code.visualstudio.com/api/commits/insider/${this.getBuildApiName(runtime)}?released=${releasedOnly}`;
-        console.log(`${chalk.gray('[build]')} fetching all builds from ${chalk.green(url)}...`);
+    private async fetchAllBuilds({ runtime, quality, flavor }: IBuildKind, releasedOnly = false): Promise<IBuild[]> {
+        const url = `https://update.code.visualstudio.com/api/commits/${quality}/${this.getBuildApiName({ runtime, quality, flavor })}?released=${releasedOnly}`;
+        LOGGER.log(`${chalk.gray('[build]')} fetching all builds from ${chalk.green(url)}...`);
         const commits = await jsonGet<Array<string>>(url);
 
-        return commits.map(commit => ({ commit, runtime }));
+        return commits.map(commit => ({ commit, runtime, quality, flavor }));
     }
 
-    private getBuildApiName(runtime: Runtime): string {
-        switch (runtime) {
-            case Runtime.WebLocal:
-            case Runtime.WebRemote:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                    case Platform.MacOSArm:
-                        return 'server-darwin-web';
-                    case Platform.LinuxX64:
-                    case Platform.LinuxArm:
-                        return 'server-linux-x64-web';
-                    case Platform.WindowsX64:
-                    case Platform.WindowsArm:
-                        return 'server-win32-x64-web';
-                }
+    private getBuildApiName({ runtime, flavor }: IBuildKind): string {
 
-            case Runtime.DesktopLocal:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                        return 'darwin';
-                    case Platform.MacOSArm:
-                        return 'darwin-arm64';
-                    case Platform.LinuxX64:
-                        return 'linux-x64';
-                    case Platform.LinuxArm:
-                        return 'linux-arm64';
-                    case Platform.WindowsX64:
-                        return 'win32-x64';
-                    case Platform.WindowsArm:
-                        return 'win32-arm64';
-                }
+        // Server
+        if (runtime === Runtime.WebLocal || runtime === Runtime.WebRemote) {
+            switch (platform) {
+                case Platform.MacOSX64:
+                case Platform.MacOSArm:
+                    return 'server-darwin-web';
+                case Platform.LinuxX64:
+                    return 'server-linux-x64-web';
+                case Platform.LinuxArm:
+                    return 'server-linux-arm64-web';
+                case Platform.WindowsX64:
+                    return 'server-win32-x64-web';
+                case Platform.WindowsArm:
+                    return 'server-win32-arm64-web';
+            }
+        }
+
+        // Desktop / CLI
+        switch (platform) {
+            case Platform.MacOSX64:
+                return flavor === Flavor.Universal ? 'darwin-universal' : 'darwin';
+            case Platform.MacOSArm:
+                return flavor === Flavor.Universal ? 'darwin-universal' : 'darwin-arm64';
+            case Platform.LinuxX64:
+                return 'linux-x64';
+            case Platform.LinuxArm:
+                return 'linux-arm64';
+            case Platform.WindowsX64:
+                return 'win32-x64';
+            case Platform.WindowsArm:
+                return 'win32-arm64';
         }
     }
 
-    async installBuild({ runtime, commit }: IBuild, options?: { forceReDownload: boolean }): Promise<void> {
-        const buildName = await this.getBuildArchiveName({ runtime, commit });
+    async installBuild({ runtime, commit, quality, flavor }: IBuild, options?: { forceReDownload: boolean }): Promise<void> {
+        const buildName = await this.getBuildArchiveName({ runtime, commit, quality, flavor });
 
-        const path = join(getBuildPath(commit), buildName);
+        const path = join(getBuildPath(commit, quality, flavor), buildName);
 
         const pathExists = await exists(path);
         if (pathExists && !options?.forceReDownload) {
-            if (LOGGER.verbose) {
-                console.log(`${chalk.gray('[build]')} using ${chalk.green(path)} for the next build to try`);
-            }
+            LOGGER.trace(`${chalk.gray('[build]')} using ${chalk.green(path)} for the next build to try`);
 
             return; // assume the build is cached
         }
 
         if (pathExists && options?.forceReDownload) {
-            console.log(`${chalk.gray('[build]')} deleting ${chalk.green(getBuildPath(commit))} and retrying download`);
-            rmSync(getBuildPath(commit), { recursive: true });
+            LOGGER.log(`${chalk.gray('[build]')} deleting ${chalk.green(getBuildPath(commit, quality, flavor))} and retrying download`);
+            rmSync(getBuildPath(commit, quality, flavor), { recursive: true });
         }
 
         // Download
-        const { url, sha256hash: expectedSHA256 } = await this.fetchBuildMeta({ runtime, commit });
-        console.log(`${chalk.gray('[build]')} downloading build from ${chalk.green(url)}...`);
+        const { url, sha256hash: expectedSHA256 } = await this.fetchBuildMeta({ runtime, commit, quality, flavor });
+        LOGGER.log(`${chalk.gray('[build]')} downloading build from ${chalk.green(url)}...`);
         await fileGet(url, path);
 
         // Validate SHA256 Checksum
@@ -157,188 +167,237 @@ class Builds {
         if (expectedSHA256 !== computedSHA256) {
             throw new Error(`${chalk.gray('[build]')} ${chalk.red('✘')} expected SHA256 checksum (${expectedSHA256}) does not match with download (${computedSHA256})`);
         } else {
-            console.log(`${chalk.gray('[build]')} ${chalk.green('✔︎')} expected SHA256 checksum matches with download`);
+            LOGGER.log(`${chalk.gray('[build]')} ${chalk.green('✔︎')} expected SHA256 checksum matches with download`);
         }
 
         // Unzip
         let destination: string;
-        if (runtime === Runtime.DesktopLocal && platform === Platform.WindowsX64 || platform === Platform.WindowsArm) {
+        if ((runtime === Runtime.DesktopLocal || runtime === Runtime.WebLocal) && flavor === Flavor.Default && (platform === Platform.WindowsX64 || platform === Platform.WindowsArm)) {
             // zip does not contain a single top level folder to use...
             destination = path.substring(0, path.lastIndexOf('.zip'));
         } else {
             // zip contains a single top level folder to use
             destination = dirname(path);
         }
-        console.log(`${chalk.gray('[build]')} unzipping build to ${chalk.green(destination)}...`);
+        LOGGER.log(`${chalk.gray('[build]')} unzipping ${chalk.green(path)} to ${chalk.green(destination)}...`);
         await unzip(path, destination);
     }
 
-    private async getBuildArchiveName({ runtime, commit }: IBuild): Promise<string> {
-        switch (runtime) {
+    private async getBuildArchiveName({ runtime, commit, quality, flavor }: IBuild): Promise<string> {
 
-            // We currently do not have ARM enabled servers
-            // so we fallback to x64 until we ship ARM.
-            case Runtime.WebLocal:
-            case Runtime.WebRemote:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                        return 'vscode-server-darwin-x64-web.zip';
-                    case Platform.MacOSArm:
-                        return 'vscode-server-darwin-arm64-web.zip';
-                    case Platform.LinuxX64:
-                        return 'vscode-server-linux-arm64-web.tar.gz';
-                    case Platform.LinuxArm:
-                        return 'vscode-server-linux-x64-web.tar.gz';
-                    case Platform.WindowsX64:
-                    case Platform.WindowsArm:
-                        return 'vscode-server-win32-x64-web.zip';
+        // Server
+        if (runtime === Runtime.WebLocal || runtime === Runtime.WebRemote) {
+            switch (platform) {
+                case Platform.MacOSX64:
+                    return 'vscode-server-darwin-x64-web.zip';
+                case Platform.MacOSArm:
+                    return 'vscode-server-darwin-arm64-web.zip';
+                case Platform.LinuxX64:
+                    return 'vscode-server-linux-x64-web.tar.gz';
+                case Platform.LinuxArm:
+                    return 'vscode-server-linux-arm64-web.tar.gz';
+                case Platform.WindowsX64:
+                    return 'vscode-server-win32-x64-web.zip';
+                case Platform.WindowsArm:
+                    return 'vscode-server-win32-arm64-web.zip';
+            }
+        }
+
+        // Desktop
+        if (flavor !== Flavor.Cli) {
+            switch (platform) {
+                case Platform.MacOSX64:
+                    return flavor === Flavor.Universal ? 'VSCode-darwin-universal.zip' : 'VSCode-darwin.zip';
+                case Platform.MacOSArm:
+                    return flavor === Flavor.Universal ? 'VSCode-darwin-universal.zip' : 'VSCode-darwin-arm64.zip';
+                case Platform.LinuxX64:
+                case Platform.LinuxArm:
+                    return (await this.fetchBuildMeta({ runtime, commit, quality, flavor })).url.split('/').pop()!; // e.g. https://az764295.vo.msecnd.net/insider/807bf598bea406dcb272a9fced54697986e87768/code-insider-x64-1639979337.tar.gz
+                case Platform.WindowsX64:
+                case Platform.WindowsArm: {
+                    const buildMeta = await this.fetchBuildMeta({ runtime, commit, quality, flavor });
+
+                    return platform === Platform.WindowsX64 ? `VSCode-win32-x64-${buildMeta.productVersion}.zip` : `VSCode-win32-arm64-${buildMeta.productVersion}.zip`;
                 }
+            }
+        }
 
-            // Every platform has its own name scheme, hilarious right?
-            // - macOS: just the name, nice! (e.g. VSCode-darwin.zip)
-            // - Linux: includes some unix timestamp (e.g. code-insider-x64-1639979337.tar.gz)
-            // - Windows: includes the version (e.g. VSCode-win32-x64-1.64.0-insider.zip)
-            case Runtime.DesktopLocal:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                        return 'VSCode-darwin.zip';
-                    case Platform.MacOSArm:
-                        return 'VSCode-darwin-arm64.zip';
-                    case Platform.LinuxX64:
-                    case Platform.LinuxArm:
-                        return (await this.fetchBuildMeta({ runtime, commit })).url.split('/').pop()!; // e.g. https://az764295.vo.msecnd.net/insider/807bf598bea406dcb272a9fced54697986e87768/code-insider-x64-1639979337.tar.gz
-                    case Platform.WindowsX64:
-                    case Platform.WindowsArm: {
-                        const buildMeta = await this.fetchBuildMeta({ runtime, commit });
-
-                        return platform === Platform.WindowsX64 ? `VSCode-win32-x64-${buildMeta.productVersion}.zip` : `VSCode-win32-arm64-${buildMeta.productVersion}.zip`;
-                    }
-                }
+        // CLI
+        switch (platform) {
+            case Platform.MacOSX64:
+                return 'vscode_cli_darwin_x64_cli.zip';
+            case Platform.MacOSArm:
+                return 'vscode_cli_darwin_arm64_cli.zip';
+            case Platform.LinuxX64:
+                return 'vscode_cli_linux_x64_cli.tar.gz';
+            case Platform.LinuxArm:
+                return 'vscode_cli_linux_arm64_cli.tar.gz';
+            case Platform.WindowsX64:
+                return 'vscode_cli_win32_x64_cli.zip';
+            case Platform.WindowsArm:
+                return 'vscode_cli_win32_arm64_cli.zip';
         }
     }
 
-    async getBuildName({ runtime, commit }: IBuild): Promise<string> {
-        switch (runtime) {
-            case Runtime.WebLocal:
-            case Runtime.WebRemote:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                        return 'vscode-server-darwin-x64-web';
-                    case Platform.MacOSArm:
-                        return 'vscode-server-darwin-arm64-web';
-                    case Platform.LinuxX64:
-                        return 'vscode-server-linux-arm64-web';
-                    case Platform.LinuxArm:
-                        return 'vscode-server-linux-x64-web';
-                    case Platform.WindowsX64:
-                    case Platform.WindowsArm:
-                        return 'vscode-server-win32-x64-web';
-                }
+    private async getBuildName({ runtime, commit, quality, flavor }: IBuild): Promise<string> {
 
-            // Here, only Windows does not play by our rules and adds the version number
-            // - Windows: includes the version (e.g. VSCode-win32-x64-1.64.0-insider)
-            case Runtime.DesktopLocal:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                    case Platform.MacOSArm:
-                        return 'Visual Studio Code - Insiders.app';
-                    case Platform.LinuxX64:
-                        return 'VSCode-linux-x64';
-                    case Platform.LinuxArm:
-                        return 'VSCode-linux-arm64';
-                    case Platform.WindowsX64:
-                    case Platform.WindowsArm: {
-                        const buildMeta = await this.fetchBuildMeta({ runtime, commit });
+        // Server
+        if (runtime === Runtime.WebLocal || runtime === Runtime.WebRemote) {
+            switch (platform) {
+                case Platform.MacOSX64:
+                    return 'vscode-server-darwin-x64-web';
+                case Platform.MacOSArm:
+                    return 'vscode-server-darwin-arm64-web';
+                case Platform.LinuxX64:
+                    return 'vscode-server-linux-x64-web';
+                case Platform.LinuxArm:
+                    return 'vscode-server-linux-arm64-web';
+                case Platform.WindowsX64:
+                    return 'vscode-server-win32-x64-web';
+                case Platform.WindowsArm:
+                    return 'vscode-server-win32-arm64-web';
+            }
+        }
 
-                        return platform === Platform.WindowsX64 ? `VSCode-win32-x64-${buildMeta.productVersion}` : `VSCode-win32-arm64-${buildMeta.productVersion}`;
-                    }
+        // Desktop
+        if (flavor !== Flavor.Cli) {
+            switch (platform) {
+                case Platform.MacOSX64:
+                case Platform.MacOSArm:
+                    return quality === 'insider' ? 'Visual Studio Code - Insiders.app' : 'Visual Studio Code.app';
+                case Platform.LinuxX64:
+                    return 'VSCode-linux-x64';
+                case Platform.LinuxArm:
+                    return 'VSCode-linux-arm64';
+                case Platform.WindowsX64:
+                case Platform.WindowsArm: {
+                    const buildMeta = await this.fetchBuildMeta({ runtime, commit, quality, flavor });
+
+                    return platform === Platform.WindowsX64 ? `VSCode-win32-x64-${buildMeta.productVersion}` : `VSCode-win32-arm64-${buildMeta.productVersion}`;
                 }
+            }
+        }
+
+        // CLI
+        return quality === 'insider' ? 'code-insiders' : 'code';
+    }
+
+    private fetchBuildMeta({ runtime, commit, quality, flavor }: IBuild): Promise<IBuildMetadata> {
+        return jsonGet<IBuildMetadata>(`https://update.code.visualstudio.com/api/versions/commit:${commit}/${this.getPlatformName({ runtime, quality, flavor })}/${quality}`);
+    }
+
+    private getPlatformName({ runtime, flavor }: IBuildKind): string {
+
+        // Server
+        if (runtime === Runtime.WebLocal || runtime === Runtime.WebRemote) {
+            switch (platform) {
+                case Platform.MacOSX64:
+                    return 'server-darwin-web';
+                case Platform.MacOSArm:
+                    return 'server-darwin-arm64-web';
+                case Platform.LinuxX64:
+                    return 'server-linux-x64-web';
+                case Platform.LinuxArm:
+                    return 'server-linux-arm64-web';
+                case Platform.WindowsX64:
+                    return 'server-win32-x64-web';
+                case Platform.WindowsArm:
+                    return 'server-win32-arm64-web';
+            }
+        }
+
+        // Desktop
+        if (flavor !== Flavor.Cli) {
+            switch (platform) {
+                case Platform.MacOSX64: {
+                    return flavor === Flavor.Universal ? 'darwin-universal' : 'darwin';
+                }
+                case Platform.MacOSArm: {
+                    return flavor === Flavor.Universal ? 'darwin-universal' : 'darwin-arm64';
+                }
+                case Platform.LinuxX64:
+                    return 'linux-x64';
+                case Platform.LinuxArm:
+                    return 'linux-arm64';
+                case Platform.WindowsX64:
+                    return 'win32-x64-archive';
+                case Platform.WindowsArm:
+                    return 'win32-arm64-archive';
+            }
+        }
+
+        // CLI
+        switch (platform) {
+            case Platform.MacOSX64:
+                return 'cli-darwin-x64';
+            case Platform.MacOSArm:
+                return 'cli-darwin-arm64';
+            case Platform.LinuxX64:
+                return 'cli-linux-x64';
+            case Platform.LinuxArm:
+                return 'cli-linux-arm64';
+            case Platform.WindowsX64:
+                return 'cli-win32-x64';
+            case Platform.WindowsArm:
+                return 'cli-win32-arm64';
         }
     }
 
-    private getPlatformName(runtime: Runtime): string {
-        switch (runtime) {
-            case Runtime.WebLocal:
-            case Runtime.WebRemote:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                        return 'server-darwin-web';
-                    case Platform.MacOSArm:
-                        return 'server-darwin-arm64-web';
-                    case Platform.LinuxX64:
-                        return 'server-linux-x64-web';
-                    case Platform.LinuxArm:
-                        return 'server-linux-arm64-web';
-                    case Platform.WindowsX64:
-                    case Platform.WindowsArm:
-                        return 'server-win32-x64-web';
-                }
+    async getBuildExecutable({ runtime, commit, quality, flavor }: IBuild): Promise<string> {
+        const buildPath = getBuildPath(commit, quality, flavor);
+        const buildName = await builds.getBuildName({ runtime, commit, quality, flavor });
 
-            case Runtime.DesktopLocal:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                        return 'darwin';
-                    case Platform.MacOSArm:
-                        return 'darwin-arm64';
-                    case Platform.LinuxX64:
-                        return 'linux-x64';
-                    case Platform.LinuxArm:
-                        return 'linux-arm64';
-                    case Platform.WindowsX64:
-                        return 'win32-x64-archive';
-                    case Platform.WindowsArm: {
-                        return 'win32-arm64-archive';
+        // Server
+        if (runtime === Runtime.WebLocal || runtime === Runtime.WebRemote) {
+            switch (platform) {
+                case Platform.MacOSX64:
+                case Platform.MacOSArm:
+                case Platform.LinuxX64:
+                case Platform.LinuxArm: {
+                    const oldLocation = join(buildPath, buildName, 'server.sh');
+                    if (await exists(oldLocation)) {
+                        return oldLocation; // only valid until 1.64.x
                     }
+
+                    return join(buildPath, buildName, 'bin', quality === 'insider' ? 'code-server-insiders' : 'code-server');
                 }
+                case Platform.WindowsX64:
+                case Platform.WindowsArm: {
+                    const oldLocation = join(buildPath, buildName, 'server.cmd');
+                    if (await exists(oldLocation)) {
+                        return oldLocation; // only valid until 1.64.x
+                    }
+
+                    return join(buildPath, buildName, buildName, 'bin', quality === 'insider' ? 'code-server-insiders.cmd' : 'code-server.cmd');
+                }
+            }
         }
-    }
 
-    private fetchBuildMeta({ runtime, commit }: IBuild): Promise<IBuildMetadata> {
-        return jsonGet<IBuildMetadata>(`https://update.code.visualstudio.com/api/versions/commit:${commit}/${this.getPlatformName(runtime)}/insider`);
-    }
+        // Desktop
+        if (flavor !== Flavor.Cli) {
+            switch (platform) {
+                case Platform.MacOSX64:
+                case Platform.MacOSArm:
+                    return join(buildPath, buildName, 'Contents', 'Resources', 'app', 'bin', 'code')
+                case Platform.LinuxX64:
+                case Platform.LinuxArm:
+                    return join(buildPath, buildName, 'bin', quality === 'insider' ? 'code-insiders' : 'code')
+                case Platform.WindowsX64:
+                case Platform.WindowsArm:
+                    return join(buildPath, buildName, 'bin', quality === 'insider' ? 'code-insiders.cmd' : 'code.cmd')
+            }
+        }
 
-    async getBuildExecutable({ runtime, commit }: IBuild): Promise<string> {
-        const buildPath = getBuildPath(commit);
-        const buildName = await builds.getBuildName({ runtime, commit });
-
-        switch (runtime) {
-            case Runtime.WebLocal:
-            case Runtime.WebRemote:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                    case Platform.MacOSArm:
-                    case Platform.LinuxX64:
-                    case Platform.LinuxArm: {
-                        const oldLocation = join(buildPath, buildName, 'server.sh');
-                        if (await exists(oldLocation)) {
-                            return oldLocation; // only valid until 1.64.x
-                        }
-
-                        return join(buildPath, buildName, 'bin', 'code-server-insiders');
-                    }
-                    case Platform.WindowsX64:
-                    case Platform.WindowsArm: {
-                        const oldLocation = join(buildPath, buildName, 'server.cmd');
-                        if (await exists(oldLocation)) {
-                            return oldLocation; // only valid until 1.64.x
-                        }
-
-                        return join(buildPath, buildName, 'bin', 'code-server-insiders.cmd');
-                    }
-                }
-
-            case Runtime.DesktopLocal:
-                switch (platform) {
-                    case Platform.MacOSX64:
-                    case Platform.MacOSArm:
-                        return join(buildPath, buildName, 'Contents', 'MacOS', 'Electron')
-                    case Platform.LinuxX64:
-                    case Platform.LinuxArm:
-                        return join(buildPath, buildName, 'code-insiders')
-                    case Platform.WindowsX64:
-                    case Platform.WindowsArm:
-                        return join(buildPath, buildName, 'Code - Insiders.exe')
-                }
+        // CLI
+        switch (platform) {
+            case Platform.MacOSX64:
+            case Platform.MacOSArm:
+            case Platform.LinuxX64:
+            case Platform.LinuxArm:
+                return join(buildPath, buildName);
+            case Platform.WindowsX64:
+            case Platform.WindowsArm:
+                return join(buildPath, `${buildName}.exe`);
         }
     }
 }
