@@ -4,8 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import chalk from 'chalk';
-import { tmpdir } from 'os';
-import { join } from 'path';
+import { createRequire } from 'node:module';
+import { tmpdir } from 'node:os';
+import { join } from 'node:path';
+
+const require = createRequire(import.meta.url);
 
 export const ROOT = join(tmpdir(), 'vscode-bisect');
 
@@ -25,12 +28,12 @@ export const DEFAULT_PERFORMANCE_FILE = join(ROOT, 'startup-perf.txt');
 export const PERFORMANCE_RUNS = 10;
 export const PERFORMANCE_RUN_TIMEOUT = 60000;
 
-export const VSCODE_DEV_URL = function (commit: string) {
+export const VSCODE_DEV_URL = function (commit: string, quality: Quality) {
     if (CONFIG.token) {
-        return `https://insiders.vscode.dev/github/microsoft/vscode/blob/main/package.json?vscode-version=${commit}`; // with auth state, we can use `github` route
+        return `https://${quality === Quality.Insider ? 'insiders.' : ''}vscode.dev/github/microsoft/vscode/blob/main/package.json?vscode-version=${commit}`; // with auth state, we can use `github` route
     }
 
-    return `https://insiders.vscode.dev/?vscode-version=${commit}`;
+    return `https://${quality === Quality.Insider ? 'insiders.' : ''}vscode.dev/?vscode-version=${commit}`;
 }
 
 export enum Platform {
@@ -59,13 +62,65 @@ export const platform = (() => {
 })();
 
 export enum Runtime {
-    WebLocal = 1,
-    WebRemote,
-    DesktopLocal
+    WebLocal = 'web-local',
+    WebRemote = 'web-remote',
+    DesktopLocal = 'desktop-local'
+}
+
+export function runtimeFromString(value: unknown): Runtime {
+    switch (value) {
+        case 'web':
+            return Runtime.WebLocal;
+        case 'vscode.dev':
+            return Runtime.WebRemote;
+        default:
+            return Runtime.DesktopLocal;
+    }
+}
+
+export enum Quality {
+    Insider = 'insider',
+    Stable = 'stable'
+}
+
+export function qualityFromString(value: unknown): Quality {
+    switch (value) {
+        case 'stable':
+            return Quality.Stable;
+        default:
+            return Quality.Insider;
+    }
+}
+
+export enum Flavor {
+    Default = 'default',
+    Universal = 'universal',
+    Cli = 'cli'
+}
+
+export function flavorFromString(value: unknown): Flavor {
+    switch (value) {
+        case 'universal':
+            return Flavor.Universal;
+        case 'cli':
+            return Flavor.Cli;
+        default:
+            return Flavor.Default;
+    }
 }
 
 export const LOGGER = {
-    verbose: false
+    verbose: false,
+    log: (message: string) => {
+        if (!isTesting() || LOGGER.verbose) {
+            console.log(message);
+        }
+    },
+    trace: (message: string) => {
+        if (LOGGER.verbose) {
+            console.log(message);
+        }
+    }
 }
 
 export const CONFIG = {
@@ -80,4 +135,13 @@ export function logTroubleshoot(): void {
 - run ${chalk.green('vscode-bisect --verbose')} for more detailed output
 - run ${chalk.green('vscode-bisect --reset')} to delete the cache folder
 - run ${chalk.green(`npm install -g ${packageJson.name}`)} to update to the latest version (your version: ${chalk.green(packageJson.version)})`);
+}
+
+let TESTING = false;
+export function setTesting(value: boolean) {
+    TESTING = value;
+}
+
+export function isTesting(): boolean {
+    return TESTING;
 }
