@@ -3,12 +3,14 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { https } from 'follow-redirects';
-import { createWriteStream, promises } from 'fs';
-import { dirname } from 'path';
-import { OutgoingHttpHeaders } from 'http';
+import followRedirects from 'follow-redirects';
+import { createWriteStream, promises } from 'node:fs';
+import { dirname } from 'node:path';
+import { OutgoingHttpHeaders } from 'node:http';
 import chalk from 'chalk';
 import ProgressBar from 'progress';
+
+const https = followRedirects.https;
 
 export function jsonGet<T>(url: string, headers?: OutgoingHttpHeaders): Promise<T> {
     return new Promise((resolve, reject) => {
@@ -43,8 +45,8 @@ export async function fileGet(url: string, path: string): Promise<void> {
             const totalSize = parseInt(res.headers['content-length']!, 10);
 
             const bar = new ProgressBar(`${chalk.gray('[fetch]')} [:bar] :percent of ${(totalSize / 1024 / 1024).toFixed(2)} MB (:rate MB/s)`, {
-                complete: '=',
-                incomplete: ' ',
+                complete: '▰',
+                incomplete: '▱',
                 width: 30,
                 total: totalSize / (1024 * 1024),
                 clear: true
